fix small bugs.

+++ generateCode/example1/single_fifo2/single/circular_fifo_lib/circular_fifo_lib.h
@@ -11,17 +11,17 @@
 	*/
 	typedef struct{
 		void* buffer;
-		size_t front;
-		size_t rear;
-		size_t capacity; // the max number of token
-		size_t token_size; // size of one token
-		size_t count;
+		int front;
+		int rear;
+		int capacity; // the max number of token
+		int token_size; // size of one token
+		int count;
 		
 	}circular_fifo;
 	
-	void init_fifo(circular_fifo* fifo_ptr, void* buf, size_t token_number, size_t token_size);
-	void read_fifo(circular_fifo* channel, void* dst, size_t number);
-	void write_fifo(circular_fifo* channel,void* src, size_t number);	
+	void init_fifo(circular_fifo* fifo_ptr, void* buf, int token_number, int token_size);
+	void read_fifo(circular_fifo* channel, void* dst, int number);
+	void write_fifo(circular_fifo* channel,void* src, int number);	
 	void PRINT(circular_fifo * fifo);			
 			
 	#endif
